pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nypl_py_utils"
version = "1.10.0"
authors = [
  { name="Aaron Friedman", email="aaronfriedman@nypl.org" },
]
description = "A package containing Python utilities for use across NYPL"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = []

[project.urls]
"Homepage" = "https://github.com/NYPL/python-utils"
"Bug Tracker" = "https://github.com/NYPL/python-utils/issues"

[project.optional-dependencies]
avro-client = [
    "nypl_py_utils[log-helper]",
    "fastavro>=1.11.1",
    "requests>=2.28.1"
]
cloudlibrary-client = [
    "nypl_py_utils[log-helper]",
    "requests>=2.28.1"
]
kinesis-client = [
    "nypl_py_utils[log-helper]",
    "boto3>=1.26.5",
    "botocore>=1.29.5"
]
kms-client = [
    "nypl_py_utils[log-helper]",
    "boto3>=1.26.5",
    "botocore>=1.29.5"
]
mysql-client = [
    "nypl_py_utils[log-helper]",
    "mysql-connector-python>=8.0.32"
]
oauth2-api-client = [
    "nypl_py_utils[log-helper]",
    "oauthlib>=3.2.2",
    "requests_oauthlib>=1.3.1"
]
postgresql-client = [
    "nypl_py_utils[log-helper]",
    "psycopg[binary]>=3.1.6"
]
redshift-client = [
    "nypl_py_utils[log-helper]",
    "botocore>=1.29.5",
    "redshift-connector>=2.0.909"
]
s3-client = [
    "nypl_py_utils[log-helper]",
    "boto3>=1.26.5",
    "botocore>=1.29.5"
]
secrets-manager-client = [
    "nypl_py_utils[log-helper]",
    "boto3>=1.26.5",
    "botocore>=1.29.5"
]
snowflake-client = [
    "nypl_py_utils[log-helper]",
    "snowflake-connector-python>=4.3.0"
]
sftp-client = [
    "nypl_py_utils[log-helper]",
    "paramiko>=3.4.1"
]
config-helper = [
    "nypl_py_utils[kms-client,log-helper]",
    "PyYAML>=6.0"
]
log-helper = [
    "structlog>=25.5.0"
]
obfuscation-helper = [
    "nypl_py_utils[log-helper]",
    "bcrypt>=4.0.1"
]
patron-data-helper = [
    "nypl_py_utils[postgresql-client,redshift-client,log-helper]>=1.1.5",
    "pandas>=2.2.2"
]
research-catalog-identifier-helper = [
    "requests>=2.28.1"
]
development = [
    "nypl_py_utils[avro-client,kinesis-client,kms-client,mysql-client,oauth2-api-client,postgresql-client,redshift-client,s3-client,secrets-manager-client,sftp-client,config-helper,log-helper,obfuscation-helper,patron-data-helper,research-catalog-identifier-helper]",
    "flake8>=6.0.0",
    "freezegun>=1.2.2",
    "mock>=4.0.3",
    "pytest==8.0",
    "pytest-mock>=3.10.0",
    "requests-mock>=1.10.0"
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q"
pythonpath = "src"
testpaths = [
    "tests"
]
